+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/API/Mempool.hs
@@ -208,6 +208,7 @@ mkMempoolEnv
       , Ledger.ledgerIx = minBound
       , Ledger.ledgerPp = nesEs ^. curPParamsEpochStateL
       , Ledger.ledgerAccount = LedgerState.esAccountState nesEs
+      , Ledger.ledgerMempool = True
       }
 
 -- | Construct a mempool state from the wider ledger state.